test: consolidate test/bench files to one sibling file per source
Mechanical move-and-merge with no test or benchmark logic changes. Route
every Test*/Benchmark* to the sibling file of the source it exercises by
function-name prefix:

- tests merged into {regextra,unmarshal,decoder}_test.go
- benches renamed/merged into {regextra,unmarshal,decoder}_bench_test.go
- bench_internal_test.go (package regextra) and bench_sanity_test.go kept
  as the package-scoped exceptions

Cross-cutting topical files (casefold, duplicate_groups, unmarshal_exclude/
optional/overflow/tag) are split by func-name prefix, converting the four
internal-package files to package regextra_test (none reference unexported
symbols). All 103 top-level Test/Example/Fuzz/Benchmark funcs preserved.

Hardening to prevent recurrence: document the one-file-per-source rule and
refresh the stale Project Structure section in CONTRIBUTING.md; add AGENTS.md
(pointer to CONTRIBUTING.md) and CLAUDE.md (pointer to AGENTS.md).

Fixes #138
