+++ CLAUDE.md
@@ -1,0 +1,2 @@
+See [AGENTS.md](AGENTS.md) for agent guidance, which points to
+[CONTRIBUTING.md](CONTRIBUTING.md) as the authoritative contributor guide.
