+++ unmarshal_test.go
@@ -3,8 +3,10 @@ package regextra_test
 import (
 	"fmt"
 	rx "github.com/jecoms/regextra"
+	"math"
 	"reflect"
 	"regexp"
+	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -1058,4 +1060,576 @@ func containsByte(s string, b byte) bool {
 		}
 	}
 	return false
+}
+
+// Tests for https://github.com/Jecoms/regextra/issues/110: the Unmarshal and
+// Decoder paths previously used two different case-folding implementations
+// (Unicode-aware strings.ToLower vs a hand-rolled ASCII lower), which could
+// disagree on non-ASCII field names. Both now use strings.EqualFold.
+//
+// Note: Go's regexp rejects non-ASCII capture group names, so divergence was
+// only possible via the *field* name side of the comparison.
+
+func TestCaseInsensitiveFallback_parity(t *testing.T) {
+	// The field name "KELVIN" below starts with U+212A (KELVIN SIGN), an
+	// uppercase Unicode letter that folds to ASCII 'k' under Unicode case
+	// folding but is untouched by ASCII-only lowering. Before the fix,
+	// Unmarshal matched it to the "kelvin" group and the Decoder did not.
+	const pattern = `(?P<kelvin>\d+)`
+	type reading struct {
+		KELVIN int
+	}
+
+	re := regexp.MustCompile(pattern)
+	var fromUnmarshal reading
+	if err := rx.Unmarshal(re, "273", &fromUnmarshal); err != nil {
+		t.Fatalf("Unmarshal: %v", err)
+	}
+
+	dec, err := rx.Compile[reading](pattern)
+	if err != nil {
+		t.Fatalf("Compile: %v", err)
+	}
+	fromDecoder, err := dec.One("273")
+	if err != nil {
+		t.Fatalf("Decoder.One: %v", err)
+	}
+
+	if fromUnmarshal.KELVIN != 273 {
+		t.Errorf("Unmarshal: KELVIN = %d, want 273", fromUnmarshal.KELVIN)
+	}
+	if fromDecoder.KELVIN != 273 {
+		t.Errorf("Decoder.One: KELVIN = %d, want 273", fromDecoder.KELVIN)
+	}
+}
+
+func TestCaseInsensitiveFallback_ascii(t *testing.T) {
+	// Plain ASCII case-insensitive fallback keeps working in both paths.
+	const pattern = `(?P<username>\w+)`
+	type rec struct {
+		UserName string
+	}
+
+	re := regexp.MustCompile(pattern)
+	var u rec
+	if err := rx.Unmarshal(re, "alice", &u); err != nil {
+		t.Fatalf("Unmarshal: %v", err)
+	}
+	if u.UserName != "alice" {
+		t.Errorf("Unmarshal: UserName = %q, want %q", u.UserName, "alice")
+	}
+
+	dec := rx.MustCompile[rec](pattern)
+	got, err := dec.One("alice")
+	if err != nil {
+		t.Fatalf("Decoder.One: %v", err)
+	}
+	if got.UserName != "alice" {
+		t.Errorf("Decoder.One: UserName = %q, want %q", got.UserName, "alice")
+	}
+}
+
+func TestUnmarshal_duplicateNames(t *testing.T) {
+	re := regexp.MustCompile(`(?:x(?P<word>a)|y(?P<word>b))`)
+	type rec struct {
+		Word string `regex:"word"`
+	}
+
+	for input, want := range map[string]string{"xa": "a", "yb": "b"} {
+		var r rec
+		if err := rx.Unmarshal(re, input, &r); err != nil {
+			t.Fatalf("Unmarshal(%q): %v", input, err)
+		}
+		if r.Word != want {
+			t.Errorf("Unmarshal(%q): Word = %q, want %q", input, r.Word, want)
+		}
+	}
+}
+
+func TestUnmarshalAll_duplicateNames(t *testing.T) {
+	re := regexp.MustCompile(`(?:x(?P<word>a)|y(?P<word>b))`)
+	type rec struct {
+		Word string `regex:"word"`
+	}
+
+	var recs []rec
+	if err := rx.UnmarshalAll(re, "xa yb xa", &recs); err != nil {
+		t.Fatalf("UnmarshalAll: %v", err)
+	}
+	want := []string{"a", "b", "a"}
+	if len(recs) != len(want) {
+		t.Fatalf("got %d matches, want %d", len(recs), len(want))
+	}
+	for i, w := range want {
+		if recs[i].Word != w {
+			t.Errorf("match %d: Word = %q, want %q", i, recs[i].Word, w)
+		}
+	}
+}
+
+// A non-participating optional group on a typed field is left at its zero value
+// by both paths; Unmarshal no longer errors trying to convert "" (issue-105
+// alignment of the Unmarshal and Decoder paths).
+func TestUnmarshalDecoder_nonParticipatingOptionalTyped(t *testing.T) {
+	const pattern = `y(?P<num>\d+)?`
+	re := regexp.MustCompile(pattern)
+	type rec struct {
+		Num int `regex:"num"`
+	}
+
+	var fromUnmarshal rec
+	if err := rx.Unmarshal(re, "y", &fromUnmarshal); err != nil {
+		t.Fatalf("Unmarshal: unexpected error %v (a non-participating optional group should leave the field zero)", err)
+	}
+	if fromUnmarshal.Num != 0 {
+		t.Errorf("Unmarshal Num = %d, want 0", fromUnmarshal.Num)
+	}
+
+	dec := rx.MustCompile[rec](pattern)
+	fromDecoder, err := dec.One("y")
+	if err != nil {
+		t.Fatalf("One: %v", err)
+	}
+	if fromDecoder.Num != 0 {
+		t.Errorf("Decoder Num = %d, want 0", fromDecoder.Num)
+	}
+}
+
+// Regression tests for https://github.com/Jecoms/regextra/issues/104:
+// an optional named group that did not participate in the match must leave
+// the field unchanged (matching Decoder behavior) instead of feeding "" to
+// the type converter and erroring.
+
+func TestUnmarshal_optionalGroupNotParticipating(t *testing.T) {
+	re := regexp.MustCompile(`(?P<name>\w+)(?: (?P<age>\d+))?`)
+
+	type person struct {
+		Name string `regex:"name"`
+		Age  int    `regex:"age"`
+	}
+
+	t.Run("absent optional group leaves field unchanged", func(t *testing.T) {
+		var p person
+		if err := rx.Unmarshal(re, "Alice", &p); err != nil {
+			t.Fatalf("Unmarshal: %v", err)
+		}
+		if p.Name != "Alice" || p.Age != 0 {
+			t.Errorf("got %+v, want {Name:Alice Age:0}", p)
+		}
+	})
+
+	t.Run("participating optional group still decodes", func(t *testing.T) {
+		var p person
+		if err := rx.Unmarshal(re, "Alice 30", &p); err != nil {
+			t.Fatalf("Unmarshal: %v", err)
+		}
+		if p.Name != "Alice" || p.Age != 30 {
+			t.Errorf("got %+v, want {Name:Alice Age:30}", p)
+		}
+	})
+
+	t.Run("default substitutes for absent optional group", func(t *testing.T) {
+		type withDefault struct {
+			Name string `regex:"name"`
+			Age  int    `regex:"age,default=18"`
+		}
+		var p withDefault
+		if err := rx.Unmarshal(re, "Alice", &p); err != nil {
+			t.Fatalf("Unmarshal: %v", err)
+		}
+		if p.Age != 18 {
+			t.Errorf("Age = %d, want default 18", p.Age)
+		}
+	})
+
+	t.Run("participating empty-span group leaves field unchanged", func(t *testing.T) {
+		// `age` participates in the match here but with a zero-length span —
+		// distinct from not participating at all, yet the contract is the
+		// same: no default, no write, no conversion error.
+		emptySpanRe := regexp.MustCompile(`(?P<name>\w+):(?P<age>\d*)`)
+		p := person{Age: 99}
+		if err := rx.Unmarshal(emptySpanRe, "Alice:", &p); err != nil {
+			t.Fatalf("Unmarshal: %v", err)
+		}
+		if p.Name != "Alice" || p.Age != 99 {
+			t.Errorf("got %+v, want {Name:Alice Age:99}", p)
+		}
+	})
+
+	t.Run("pre-populated field survives absent group", func(t *testing.T) {
+		p := person{Age: 99}
+		if err := rx.Unmarshal(re, "Alice", &p); err != nil {
+			t.Fatalf("Unmarshal: %v", err)
+		}
+		if p.Age != 99 {
+			t.Errorf("Age = %d, want 99 (unchanged)", p.Age)
+		}
+	})
+
+	t.Run("UnmarshalAll skips absent optional groups per match", func(t *testing.T) {
+		var people []person
+		if err := rx.UnmarshalAll(re, "Alice 30, Bob", &people); err != nil {
+			t.Fatalf("UnmarshalAll: %v", err)
+		}
+		if len(people) != 2 {
+			t.Fatalf("got %d matches, want 2", len(people))
+		}
+		if people[0].Age != 30 || people[1].Age != 0 {
+			t.Errorf("got %+v", people)
+		}
+	})
+}
+
+// Unmarshal and Decoder.One must agree on optional-group semantics.
+func TestUnmarshal_decoderParityOnOptionalGroups(t *testing.T) {
+	const pattern = `(?P<name>\w+)(?: (?P<age>\d+))?`
+	type person struct {
+		Name string `regex:"name"`
+		Age  int    `regex:"age"`
+	}
+
+	re := regexp.MustCompile(pattern)
+	dec := rx.MustCompile[person](pattern)
+
+	for _, input := range []string{"Alice", "Alice 30"} {
+		var fromUnmarshal person
+		if err := rx.Unmarshal(re, input, &fromUnmarshal); err != nil {
+			t.Fatalf("Unmarshal(%q): %v", input, err)
+		}
+		fromDecoder, err := dec.One(input)
+		if err != nil {
+			t.Fatalf("Decoder.One(%q): %v", input, err)
+		}
+		if fromUnmarshal != fromDecoder {
+			t.Errorf("input %q: Unmarshal=%+v Decoder.One=%+v — paths disagree", input, fromUnmarshal, fromDecoder)
+		}
+	}
+
+	// Parity must also hold for the participating empty-span case — the novel
+	// behavior this PR fixes (#104), which the optional-group pattern above
+	// can't produce (its `age` group either participates with digits or not at
+	// all). On the empty-span input both paths must skip the field identically
+	// rather than one erroring on `""`.
+	const emptySpanPattern = `(?P<name>\w+):(?P<age>\d*)`
+	emptyRe := regexp.MustCompile(emptySpanPattern)
+	emptyDec := rx.MustCompile[person](emptySpanPattern)
+	var fromUnmarshal person
+	if err := rx.Unmarshal(emptyRe, "Alice:", &fromUnmarshal); err != nil {
+		t.Fatalf("Unmarshal(empty-span): %v", err)
+	}
+	fromDecoder, err := emptyDec.One("Alice:")
+	if err != nil {
+		t.Fatalf("Decoder.One(empty-span): %v", err)
+	}
+	if fromUnmarshal != fromDecoder {
+		t.Errorf("empty-span: Unmarshal=%+v Decoder.One=%+v — paths disagree", fromUnmarshal, fromDecoder)
+	}
+}
+
+// Regression tests for https://github.com/Jecoms/regextra/issues/103:
+// values out of range for narrow numeric fields must error instead of
+// silently truncating (e.g. "300" into int8 previously yielded 44, nil).
+
+func TestUnmarshal_intOverflow(t *testing.T) {
+	re := regexp.MustCompile(`(?P<n>-?\d+)`)
+
+	t.Run("int8 overflow errors", func(t *testing.T) {
+		var dst struct {
+			N int8 `regex:"n"`
+		}
+		err := rx.Unmarshal(re, "300", &dst)
+		if err == nil {
+			t.Fatalf("expected overflow error, got nil (N = %d)", dst.N)
+		}
+		if !strings.Contains(err.Error(), "cannot convert") {
+			t.Errorf("unexpected error: %v", err)
+		}
+	})
+
+	t.Run("int8 underflow errors", func(t *testing.T) {
+		var dst struct {
+			N int8 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, "-129", &dst); err == nil {
+			t.Fatalf("expected underflow error, got nil (N = %d)", dst.N)
+		}
+	})
+
+	t.Run("int8 bounds fit", func(t *testing.T) {
+		var dst struct {
+			Lo int8 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, "127", &dst); err != nil {
+			t.Fatalf("127 should fit in int8: %v", err)
+		}
+		if dst.Lo != math.MaxInt8 {
+			t.Errorf("Lo = %d, want 127", dst.Lo)
+		}
+		if err := rx.Unmarshal(re, "-128", &dst); err != nil {
+			t.Fatalf("-128 should fit in int8: %v", err)
+		}
+		if dst.Lo != math.MinInt8 {
+			t.Errorf("Lo = %d, want -128", dst.Lo)
+		}
+	})
+
+	t.Run("int16 and int32 overflow errors", func(t *testing.T) {
+		var dst16 struct {
+			N int16 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, "40000", &dst16); err == nil {
+			t.Errorf("expected int16 overflow error, got nil (N = %d)", dst16.N)
+		}
+		var dst32 struct {
+			N int32 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, strconv.FormatInt(math.MaxInt32+1, 10), &dst32); err == nil {
+			t.Errorf("expected int32 overflow error, got nil (N = %d)", dst32.N)
+		}
+	})
+}
+
+func TestUnmarshal_uintOverflow(t *testing.T) {
+	re := regexp.MustCompile(`(?P<n>\d+)`)
+
+	t.Run("uint8 overflow errors", func(t *testing.T) {
+		var dst struct {
+			N uint8 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, "256", &dst); err == nil {
+			t.Fatalf("expected overflow error, got nil (N = %d)", dst.N)
+		}
+	})
+
+	t.Run("uint8 max fits", func(t *testing.T) {
+		var dst struct {
+			N uint8 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, "255", &dst); err != nil {
+			t.Fatalf("255 should fit in uint8: %v", err)
+		}
+		if dst.N != math.MaxUint8 {
+			t.Errorf("N = %d, want 255", dst.N)
+		}
+	})
+}
+
+func TestUnmarshal_floatOverflow(t *testing.T) {
+	re := regexp.MustCompile(`(?P<n>[\d.e+]+)`)
+
+	t.Run("float32 overflow errors", func(t *testing.T) {
+		var dst struct {
+			N float32 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, "1e308", &dst); err == nil {
+			t.Fatalf("expected overflow error, got nil (N = %g)", dst.N)
+		}
+	})
+
+	t.Run("float32 in range fits", func(t *testing.T) {
+		var dst struct {
+			N float32 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, "3.5", &dst); err != nil {
+			t.Fatalf("3.5 should fit in float32: %v", err)
+		}
+		if dst.N != 3.5 {
+			t.Errorf("N = %g, want 3.5", dst.N)
+		}
+	})
+
+	t.Run("float64 still accepts large values", func(t *testing.T) {
+		var dst struct {
+			N float64 `regex:"n"`
+		}
+		if err := rx.Unmarshal(re, "1e308", &dst); err != nil {
+			t.Fatalf("1e308 fits in float64: %v", err)
+		}
+	})
+}
+
+// Issue #106: `regex:"-"` must exclude a field entirely, even when a declared
+// group shares the field's name (the trap the old name-fallback created). An
+// absent tag (`regex:""`) must still fall back to the field name — the guard
+// against over-skipping.
+
+func TestUnmarshal_dashExcludesField(t *testing.T) {
+	type Person struct {
+		Name string `regex:"name"`
+		Age  string `regex:"-"` // excluded, despite a declared "age" group
+	}
+	re := regexp.MustCompile(`(?P<name>\w+) is (?P<age>\d+)`)
+
+	var p Person
+	if err := rx.Unmarshal(re, "Alice is 30", &p); err != nil {
+		t.Fatalf("Unmarshal() error = %v", err)
+	}
+	if p.Name != "Alice" {
+		t.Errorf("Name = %q, want %q", p.Name, "Alice")
+	}
+	if p.Age != "" {
+		t.Errorf("Age = %q, want it left at zero value (excluded by regex:\"-\")", p.Age)
+	}
+}
+
+func TestUnmarshalAll_dashExcludesField(t *testing.T) {
+	type Person struct {
+		Name string `regex:"name"`
+		Age  string `regex:"-"`
+	}
+	re := regexp.MustCompile(`(?P<name>\w+) is (?P<age>\d+)`)
+
+	var people []Person
+	if err := rx.UnmarshalAll(re, "Alice is 30 Bob is 25", &people); err != nil {
+		t.Fatalf("UnmarshalAll() error = %v", err)
+	}
+	if len(people) != 2 {
+		t.Fatalf("got %d people, want 2", len(people))
+	}
+	for i, p := range people {
+		if p.Age != "" {
+			t.Errorf("people[%d].Age = %q, want it left at zero value", i, p.Age)
+		}
+	}
+}
+
+func TestUnmarshal_emptyTagStillFallsBackToFieldName(t *testing.T) {
+	// Guard against over-skipping: an absent tag must still match by field name.
+	type Person struct {
+		Name string
+		Age  string `regex:""`
+	}
+	re := regexp.MustCompile(`(?P<name>\w+) is (?P<age>\d+)`)
+
+	var p Person
+	if err := rx.Unmarshal(re, "Alice is 30", &p); err != nil {
+		t.Fatalf("Unmarshal() error = %v", err)
+	}
+	if p.Age != "30" {
+		t.Errorf("Age = %q, want %q (regex:\"\" should fall back to field name)", p.Age, "30")
+	}
+}
+
+// Only the bare `-` excludes: a leading `-` followed by options parses `-` as
+// the group name, which matches no group, so the field falls through to its
+// `default`. The documented boundary in three doc locations (parseFieldTag
+// godoc, README, CHANGELOG), guarded here so it can't silently regress.
+func TestUnmarshal_dashWithOptionsIsNotExcluded(t *testing.T) {
+	type Person struct {
+		Name string `regex:"name"`
+		Age  string `regex:"-,default=unknown"` // "-" is a group name, not an exclude
+	}
+	re := regexp.MustCompile(`(?P<name>\w+) is (?P<age>\d+)`)
+
+	var p Person
+	if err := rx.Unmarshal(re, "Alice is 30", &p); err != nil {
+		t.Fatalf("Unmarshal() error = %v", err)
+	}
+	if p.Age != "unknown" {
+		t.Errorf("Age = %q, want %q (regex:\"-,default=x\" is not excluded; \"-\" matches no group, so default applies)", p.Age, "unknown")
+	}
+}
+
+// Issue #113: pin the two tag-grammar forward-compatibility rules that the
+// package doc (regextra.go "Tag grammar") declares part of the v1 contract,
+// exercised through the public Unmarshal/Decoder surface rather than the
+// unexported parser:
+//
+//   - Unknown key=value pairs are accepted, not rejected, so a future minor
+//     release can recognize new option keys without a parser change.
+//   - Lone tokens (no '=') are silently ignored, reserving that slot for
+//     future flag-style options (e.g. `required`).
+//
+// Each rule is observable only as a no-op today: the extra option carries no
+// meaning, so the field resolves exactly as if it were absent. Pairing the
+// forward-compat piece with a `default` on an absent group makes that visible
+// — the default still applies, proving the parser neither errored on nor was
+// derailed by the extra piece. (Whether an unknown key is retained in the
+// internal option map is not observable through the public API, so the tests
+// pin only that such keys are accepted, not that they are stored.)
+
+func TestUnmarshal_unknownOptionKeyIsAccepted(t *testing.T) {
+	// Forward-compat rule 1: an unrecognized key=value must not be rejected.
+	// `future` has no meaning today, so the field matches its group exactly as
+	// if the option were absent.
+	type Person struct {
+		Name string `regex:"name,future=42"`
+	}
+	re := regexp.MustCompile(`(?P<name>\w+)`)
+
+	var p Person
+	if err := rx.Unmarshal(re, "Alice", &p); err != nil {
+		t.Fatalf("Unmarshal() error = %v, want nil (unknown option keys must be accepted, not rejected)", err)
+	}
+	if p.Name != "Alice" {
+		t.Errorf("Name = %q, want %q", p.Name, "Alice")
+	}
+}
+
+func TestUnmarshal_loneTokenIsIgnored(t *testing.T) {
+	// Forward-compat rule 2: a lone token (no '=') is silently ignored today,
+	// reserving the slot for future flag-style options. `required` has no
+	// meaning yet, so it neither errors nor blocks the `default` that follows
+	// it — the "role" group is absent, so the field takes the default.
+	type Person struct {
+		Role string `regex:"role,required,default=guest"`
+	}
+	re := regexp.MustCompile(`(?P<name>\w+)`) // no "role" group
+
+	var p Person
+	if err := rx.Unmarshal(re, "Alice", &p); err != nil {
+		t.Fatalf("Unmarshal() error = %v, want nil (a lone token must be ignored)", err)
+	}
+	if p.Role != "guest" {
+		t.Errorf("Role = %q, want %q (lone token ignored; default still applies)", p.Role, "guest")
+	}
+}
+
+func TestUnmarshal_emptyOptionPieceIsSkipped(t *testing.T) {
+	// An empty option piece is ignored without disturbing the options around
+	// it: the `default` still resolves when the group is absent. The behavior
+	// holds the same way wherever the empty piece falls, so these positional
+	// variants share subtests (struct tags are compile-time literals, so each
+	// needs its own struct rather than a value table). An empty piece has no
+	// '=', so the parser drops it via the same path as a lone token.
+	re := regexp.MustCompile(`(?P<name>\w+)`) // no "role" group; default fires
+
+	t.Run("doubled comma", func(t *testing.T) {
+		type Person struct {
+			Role string `regex:"role,,default=guest"`
+		}
+		var p Person
+		if err := rx.Unmarshal(re, "Alice", &p); err != nil {
+			t.Fatalf("Unmarshal() error = %v, want nil", err)
+		}
+		if p.Role != "guest" {
+			t.Errorf("Role = %q, want %q (empty piece skipped; default still applies)", p.Role, "guest")
+		}
+	})
+
+	t.Run("trailing comma", func(t *testing.T) {
+		type Person struct {
+			Role string `regex:"role,default=guest,"`
+		}
+		var p Person
+		if err := rx.Unmarshal(re, "Alice", &p); err != nil {
+			t.Fatalf("Unmarshal() error = %v, want nil", err)
+		}
+		if p.Role != "guest" {
+			t.Errorf("Role = %q, want %q (trailing empty piece skipped)", p.Role, "guest")
+		}
+	})
+
+	t.Run("whitespace-only piece", func(t *testing.T) {
+		type Person struct {
+			Role string `regex:"role, ,default=guest"`
+		}
+		var p Person
+		if err := rx.Unmarshal(re, "Alice", &p); err != nil {
+			t.Fatalf("Unmarshal() error = %v, want nil", err)
+		}
+		if p.Role != "guest" {
+			t.Errorf("Role = %q, want %q (whitespace-only piece skipped)", p.Role, "guest")
+		}
+	})
 }
